split `types/configurations.ts` into multiple files in `configurations` sub-module
