+++ src/app-support/types/index.ts
@@ -2,4 +2,4 @@ export * from "./apps";
 export * from "./observables";
 export * from "./payload";
 export * from "./known-chains";
-export * as Config from "./configurations";
+export * from "./configurations";
